Added engines directory folder and rename GameRender to RenderEgine
